CDD-2169: move landing page model to home app
