+++ cms/home/managers/__init__.py
@@ -1,0 +1,2 @@
+from .homepage import HomePageManager, EXPECTED_HOME_PAGE_SLUG
+from .landingpage import LandingPageManager, EXPECTED_LANDING_PAGE_SLUG
