+++ cms/home/models/__init__.py
@@ -1,0 +1,2 @@
+from .homepage import HomePage, HomePageRelatedLink, UKHSARootPage
+from .landingpage import LandingPage
