+++ cms/home/models/landingpage.py
@@ -5,30 +5,22 @@ from wagtail.api import APIField
 from wagtail.fields import RichTextField
 from wagtail.models import Orderable, Page
 
-from cms.common.models import AVAILABLE_RICH_TEXT_FEATURES, MAXIMUM_URL_FIELD_LENGTH
+from cms.common.models import MAXIMUM_URL_FIELD_LENGTH
 from cms.dashboard.enums import (
     DEFAULT_RELATED_LINKS_LAYOUT_FIELD_LENGTH,
     RelatedLinksLayoutEnum,
 )
 from cms.dashboard.models import UKHSAPage
 from cms.dynamic_content import help_texts
-from cms.dynamic_content.access import ALLOWABLE_BODY_CONTENT
-from cms.home.managers import HomePageManager
+from cms.dynamic_content.access import ALLOWABLE_BODY_CONTENT_SECTION_LINK
+from cms.home.managers import LandingPageManager
 
 
-class UKHSARootPage(Page):
+class LandingPage(UKHSAPage):
+    is_creatable = True
     max_count = 1
-
-
-class HomePage(UKHSAPage):
-    max_count = 1
-    page_description = RichTextField(
-        features=AVAILABLE_RICH_TEXT_FEATURES,
-        blank=True,
-        null=True,
-        help_text=help_texts.PAGE_DESCRIPTION_FIELD,
-    )
-    body = ALLOWABLE_BODY_CONTENT
+    sub_title = models.CharField(max_length=255)
+    body = ALLOWABLE_BODY_CONTENT_SECTION_LINK
     related_links_layout = models.CharField(
         verbose_name="Layout",
         help_text=help_texts.RELATED_LINKS_LAYOUT_FIELD,
@@ -37,27 +29,23 @@ class HomePage(UKHSAPage):
         choices=RelatedLinksLayoutEnum.choices(),
     )
 
-    content_panels = Page.content_panels + [
-        FieldPanel("page_description"),
-        FieldPanel("body"),
-    ]
+    content_panels = Page.content_panels + [FieldPanel("sub_title"), FieldPanel("body")]
 
     sidebar_content_panels = [
         FieldPanel("related_links_layout"),
         InlinePanel("related_links", heading="Related links", label="Related link"),
     ]
 
-    # Sets which fields to expose on the API
     api_fields = UKHSAPage.api_fields + [
-        APIField("page_description"),
+        APIField("title"),
+        APIField("sub_title"),
         APIField("body"),
         APIField("related_links_layout"),
         APIField("related_links"),
-        APIField("last_published_at"),
         APIField("search_description"),
+        APIField("last_published_at"),
     ]
 
-    # Tabs to position at the top of the view
     edit_handler = TabbedInterface(
         [
             ObjectList(content_panels, heading="Content"),
@@ -66,11 +54,11 @@ class HomePage(UKHSAPage):
         ]
     )
 
-    objects = HomePageManager()
+    objects = LandingPageManager()
 
     @classmethod
-    def is_previewable(cls) -> bool:
-        """Returns False. Since this is a headless CMS the preview panel is not supported"""
+    def is_previewable(cls):
+        """Returns False. This is a headline CMS, preview panel is not supported ."""
         return False
 
     def get_url_parts(self, request=None) -> tuple[int, str, str]:
@@ -97,7 +85,7 @@ class HomePage(UKHSAPage):
                 3) The path of the current page.
                     This always returns an empty string
                     e.g. `""`
-                    This is because the homepage is assumed
+                    This is because the landing page is assumed
                     to be at the root of the visible page tree.
 
         """
@@ -106,9 +94,9 @@ class HomePage(UKHSAPage):
         return site_id, root_url, page_path
 
 
-class HomePageRelatedLink(Orderable):
+class LandingPageRelatedLink(Orderable):
     page = ParentalKey(
-        HomePage, on_delete=models.SET_NULL, null=True, related_name="related_links"
+        LandingPage, on_delete=models.SET_NULL, null=True, related_name="related_links"
     )
     title = models.CharField(max_length=255)
     url = models.URLField(verbose_name="URL", max_length=MAXIMUM_URL_FIELD_LENGTH)
