--- cms/landing_page/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class LandingPageConfig(AppConfig):
-    default_auto_fields = "django.db.models.BigAutoField"
-    name = "cms.landing_page"
